renames package catalog => catalogs
