+++ src/dimcat/data/catalogs/base.py
