+++ src/dimcat/steps/loaders/base.py
@@ -23,7 +23,7 @@ import marshmallow as mm
 import pandas as pd
 from dimcat.base import FriendlyEnum, get_setting
 from dimcat.data.base import Data
-from dimcat.data.catalog.base import DimcatCatalog
+from dimcat.data.catalogs.base import DimcatCatalog
 from dimcat.data.dataset.base import Dataset
 from dimcat.data.package.base import Package, PathPackage
 from dimcat.data.package.dc import DimcatPackage
